Use event server URL for port, and vice versa
If event server URL is given with no explicit port, use the port from
the URL as the event server port. If a port and URL are given and the
URL doesn't include a port, use the explicit port. If both a port and a
URL port are provided, use the explicit port for the event server.